refactor: finalize import updates for relocated sections
