+++ src/app/(app)/download/page.tsx
@@ -1,9 +1,10 @@
-import Download from "@/sections/Download";
+import Download from "@/components/sections/Download";
 
 export default function DownloadPage() {
   return (
-    <main className="bg-transparent">
+    <main className="bg-text-color flex flex-col">
       <Download />
+      <div className="h-50 bg-white w-full" />
     </main>
   );
 }
